+++ packages/app-backend/src/vuex.js
@@ -3,9 +3,8 @@ import { cloneDeep } from 'lodash-es'
 import { stringify, cloneVueData } from './utils'
 import { set, get, parse } from '@utils/util'
 import SharedData from '@utils/shared-data'
-import { getCatchedGetters } from '@vue-devtools/shared-utils'
 import clone from './clone'
-import { debounce } from './utils'
+import { debounce, getCatchedGetters } from './utils'
 import { bridge as exBridge, api } from './bridge'
 
 const isProd = process.env.NODE_ENV === 'production'
