--- packages/shared-utils/src/details/custom.ts
@@ -1,52 +1,0 @@
-export function getCustomRouterDetails(router) {
-  return {
-    _custom: {
-      type: 'router',
-      display: 'VueRouter',
-      value: {
-        options: router.options,
-        currentRoute: router.currentRoute,
-      },
-      fields: {
-        abstract: true,
-      },
-    },
-  }
-}
-
-export function getCatchedGetters(store) {
-  const getters = {}
-
-  const origGetters = store.getters || {}
-  const keys = Object.keys(origGetters)
-  for (let i = 0; i < keys.length; i++) {
-    const key = keys[i]
-    Object.defineProperty(getters, key, {
-      enumerable: true,
-      get: () => {
-        try {
-          return origGetters[key]
-        } catch (e) {
-          return e
-        }
-      },
-    })
-  }
-
-  return getters
-}
-export function getCustomStoreDetails(store) {
-  return {
-    _custom: {
-      type: 'store',
-      display: 'Store',
-      value: {
-        state: store.state,
-        getters: getCatchedGetters(store),
-      },
-      fields: {
-        abstract: true,
-      },
-    },
-  }
-}
