+++ packages/shared-utils/src/index.js
@@ -1,2 +1,1 @@
 export * from './api'
-export { getCatchedGetters } from './details/custom'
